resume-parser-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyresparser==1.0.6
spacy==3.6.1
docx2txt==0.8
pdfminer.six==20221105
nltk==3.8.1
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==2.3.7